pyproject.toml
[project]
name = "sipet"
version = "0.1.0"
description = ""
authors = [
    {name = "jalm2035gmail",email = "jalm2035@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.10,<4"
dependencies = [
    "fastapi (>=0.128.7,<0.129.0)",
    "uvicorn (>=0.40.0,<0.41.0)",
    "sqlalchemy (>=2.0.46,<3.0.0)",
    "alembic (>=1.18.4,<2.0.0)",
    "psycopg2-binary (>=2.9.11,<3.0.0)",
    "pydantic (>=2.12.5,<3.0.0)",
    "python-jose[cryptography] (>=3.5.0,<4.0.0)",
    "passlib[bcrypt] (>=1.7.4,<2.0.0)",
    "python-multipart (>=0.0.22,<0.0.23)",
    "openpyxl (>=3.1.5,<4.0.0)",
    "reportlab (>=4.4.9,<5.0.0)",
    "celery (>=5.6.2,<6.0.0)",
    "redis (>=7.1.1,<8.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
